requirements.txt
Django==2.2.9
django-rest==0.0.1
djangorestframework==3.11.0
psycopg2==2.8.4
python-dotenv==0.10.5
pytz==2019.3
sqlparse==0.3.0
